Inlined Abbreviation button from JournalAbbreviationButton to EntryEditor
